chore: move audio class
