+++ core/src/cn/harryh/arkpets/audio/AudioManager.java
@@ -1,4 +1,4 @@
-package cn.harryh.arkpets.utils;
+package cn.harryh.arkpets.audio;
 
 import cn.harryh.arkpets.concurrent.SocketClient;
